Removed the redundant degredation of vertex coordinates
There is no reason to reduce the resolution for the x,y,z coordinates by the `x=int(x*10000)/10000` expression. My guess is that this was done to find duplicate vertices. But the worst case is that there will be a double vertex match, which really is not that bad as the only problem will then be that the file will grow a bit.